new approach, first approach was recursively calling system summary endpoint (because ato endpoint calls it)
